changed property from int to NSInteger. Removed unnecessary getters and setters for property since they are automatically synthesized.

+++ EmailFlow2/EmailListViewController.h
@@ -43,5 +43,6 @@
 
 @property (nonatomic,strong) UIColor *blueAppColor, *lightGrayAppColor, *mediumGrayAppColor, *grayBackgroundAppColor;
 @property (nonatomic,strong) NSIndexPath *currentlySelectedListCellPath;
-@property (nonatomic, assign) int currentlySelectedListCell;
+@property (nonatomic, assign) NSInteger currentlySelectedListCell;
+
 @end
